Refactor Case Transfer controller code
